requirements.txt
numpy>=1.24
matplotlib>=3.7
scipy>=1.10
networkx>=3.0
scikit-learn>=1.2
pyyaml>=6.0
